[#230] Change SingleItemCollection file structure
